requirements.txt
flask==2.0.1
xrpl-py==2.0.0
aiohttp==3.8.1
python-dotenv==0.19.0
